fix: keyboard logic
